Replace get_range_raw with fold_range_future/wait
This has been bugging me for awhile in that we're not properly handling
the return value from the range future. The `more` return value in
particular was ignored.

Rather than force people to have to repeatedly reimplement this I've
changed `get_range_raw` with a pair of functions `fold_range_future` and
`fold_range_wait`. This allows us the benefit of the async approach
while also ensuring we're not dealing with partial ranges being returned
to users.