Add a dev container configuration for VS Code (#27)
This creates a development environment with a FoundationDB server and
an erlfdb client in two containers, sharing a network through Docker
Compose.

It uses the FDB image published to Docker Hub for the FDB container, and
downloads the FDB client packages from foundationdb.org to provide the
development headers and libraries. Once the Docker Compose setup is
created, VS Code executes the `create_cluster_file.bash` script to write
down a cluster file containing the IP address in the compose network
where the FDB service can be found, and initializes the FDB server with
a new database.

The use of an external FDB container allows the erlfdb container to be
re-created as needed without losing the underlying data in the FDB
container.