Simplify config by using default fdb.cluster
A bit obvious in hindsight. We don't need to be passing around custom
overrides for the fdb.cluster file location; all we really need to do
is just trust FDB to figure it out on each platform. To do this we allow
for setting the test_cluster_file env to `system_default` instead of a
path to a specific cluster file.