Avoid reserved keys in tests
Prefix keys with a 0 byte to avoid the /xFF space. I suppose we could
apply this same treatment for keys and values as is being done elsewhere
in the test suite, but there's no such restriction on value encoding,
and I think it's sometimes helpful to reinforce the idea that keys have
special requirements that do not carry over to values.

Closes #46